wallet-connection-with-key-management file created
