+++ lib/wallet/index.js
@@ -19,9 +19,11 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.wallet_injected_interface = exports.ConnectedWalletAccountRedirect = exports.WalletConnectionRedirect = exports.WalletConnectionInjectedWithLocalFckManagement = exports.WalletConnectionInjected = exports.WalletConnectionWithKeyManagement = void 0;
+exports.wallet_injected_interface = exports.ConnectedWalletAccountRedirect = exports.WalletConnectionRedirect = exports.WalletConnectionInjectedWithLocalFckManagement = exports.WalletConnectionInjected = exports.WalletConnectionWithKeyManagement = exports.WalletConnection = exports.WalletConnectionType = exports.createWalletConnection = void 0;
 const wallet_connection_1 = require("./wallet-connection");
-Object.defineProperty(exports, "WalletConnectionWithKeyManagement", { enumerable: true, get: function () { return wallet_connection_1.WalletConnectionWithKeyManagement; } });
+Object.defineProperty(exports, "WalletConnection", { enumerable: true, get: function () { return wallet_connection_1.WalletConnection; } });
+const wallet_connection_with_key_management_1 = require("./wallet-connection-with-key-management");
+Object.defineProperty(exports, "WalletConnectionWithKeyManagement", { enumerable: true, get: function () { return wallet_connection_with_key_management_1.WalletConnectionWithKeyManagement; } });
 const wallet_connection_injected_1 = require("./injected/wallet-connection-injected");
 Object.defineProperty(exports, "WalletConnectionInjected", { enumerable: true, get: function () { return wallet_connection_injected_1.WalletConnectionInjected; } });
 const wallet_connection_injected_with_local_fck_management_1 = require("./injected/wallet-connection-injected-with-local-fck-management");
@@ -29,5 +31,8 @@ Object.defineProperty(exports, "WalletConnectionInjectedWithLocalFckManagement",
 const wallet_connection_redirect_1 = require("./redirect/wallet-connection-redirect");
 Object.defineProperty(exports, "WalletConnectionRedirect", { enumerable: true, get: function () { return wallet_connection_redirect_1.WalletConnectionRedirect; } });
 Object.defineProperty(exports, "ConnectedWalletAccountRedirect", { enumerable: true, get: function () { return wallet_connection_redirect_1.ConnectedWalletAccountRedirect; } });
+const wallet_connection_creator_1 = require("./wallet-connection-creator");
+Object.defineProperty(exports, "createWalletConnection", { enumerable: true, get: function () { return wallet_connection_creator_1.createWalletConnection; } });
+Object.defineProperty(exports, "WalletConnectionType", { enumerable: true, get: function () { return wallet_connection_creator_1.WalletConnectionType; } });
 const wallet_injected_interface = __importStar(require("./injected/wallet-injected-interface"));
 exports.wallet_injected_interface = wallet_injected_interface;
