+++ lib/wallet/injected/wallet-connection-injected-with-local-fck-management.js
@@ -1,8 +1,8 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.ConnectedWalletAccountInjected = exports.WalletConnectionInjectedWithLocalFckManagement = void 0;
-const wallet_connection_1 = require("../wallet-connection");
-class WalletConnectionInjectedWithLocalFckManagement extends wallet_connection_1.WalletConnectionWithKeyManagement {
+exports.ConnectedWalletAccountInjectedWithKeyManagement = exports.WalletConnectionInjectedWithLocalFckManagement = void 0;
+const wallet_connection_with_key_management_1 = require("../wallet-connection-with-key-management");
+class WalletConnectionInjectedWithLocalFckManagement extends wallet_connection_with_key_management_1.WalletConnectionWithKeyManagement {
     constructor(near, appKeyPrefix, walletName) {
         super(near, appKeyPrefix);
         this._walletName = walletName;
@@ -15,12 +15,12 @@ class WalletConnectionInjectedWithLocalFckManagement extends wallet_connection_1
     }
     account() {
         if (!this._connectedAccount) {
-            this._connectedAccount = new ConnectedWalletAccountInjected(this, this._near.connection, this._authData.accountId);
+            this._connectedAccount = new ConnectedWalletAccountInjectedWithKeyManagement(this, this._near.connection, this._authData.accountId);
         }
         return this._connectedAccount;
     }
 }
 exports.WalletConnectionInjectedWithLocalFckManagement = WalletConnectionInjectedWithLocalFckManagement;
-class ConnectedWalletAccountInjected extends wallet_connection_1.ConnectedWalletAccount {
+class ConnectedWalletAccountInjectedWithKeyManagement extends wallet_connection_with_key_management_1.ConnectedWalletAccountWithKeyManagement {
 }
-exports.ConnectedWalletAccountInjected = ConnectedWalletAccountInjected;
+exports.ConnectedWalletAccountInjectedWithKeyManagement = ConnectedWalletAccountInjectedWithKeyManagement;
