+++ lib/wallet/injected/wallet-connection-injected.d.ts
@@ -1,7 +1,7 @@
-import { RequestSignTransactionsOptions, SignInOptions, WalletConnectionBase } from "../wallet-connection";
+import { RequestSignTransactionsOptions, SignInOptions, WalletConnection } from "../wallet-connection";
 import { Near } from '../../near';
 import { Account } from '../../account';
-export declare class WalletConnectionInjected extends WalletConnectionBase {
+export declare class WalletConnectionInjected extends WalletConnection {
     _walletName: string;
     constructor(near: Near, appKeyPrefix: string | null, walletName: string);
     requestSignTransactions({ transactions, meta, callbackUrl }: RequestSignTransactionsOptions): Promise<void>;
