+++ lib/wallet/injected/wallet-connection-injected.js
@@ -2,7 +2,7 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.WalletConnectionInjected = void 0;
 const wallet_connection_1 = require("../wallet-connection");
-class WalletConnectionInjected extends wallet_connection_1.WalletConnectionBase {
+class WalletConnectionInjected extends wallet_connection_1.WalletConnection {
     constructor(near, appKeyPrefix, walletName) {
         super(near, appKeyPrefix);
         this._walletName = walletName;
