+++ lib/wallet/redirect/wallet-connection-redirect.js
@@ -1,8 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.ConnectedWalletAccountRedirect = exports.WalletConnectionRedirect = void 0;
-const index_1 = require("../index");
-const wallet_connection_1 = require("../wallet-connection");
+const wallet_connection_with_key_management_1 = require("../wallet-connection-with-key-management");
 const utils_1 = require("../../utils");
 const borsh_1 = require("borsh");
 const transaction_1 = require("../../transaction");
@@ -27,7 +26,7 @@ const PENDING_ACCESS_KEY_PREFIX = 'pending_key'; // browser storage key for a pe
  * if(!walletConnection.isSingnedIn()) return walletConnection.requestSignIn()
  * ```
  */
-class WalletConnectionRedirect extends index_1.WalletConnectionWithKeyManagement {
+class WalletConnectionRedirect extends wallet_connection_with_key_management_1.WalletConnectionWithKeyManagement {
     /**
      * @param {string} walletBaseUrl NEAR wallet URL, used to redirect users to their wallet for signing and sending transactions
     */
@@ -139,7 +138,7 @@ exports.WalletConnectionRedirect = WalletConnectionRedirect;
 /**
  * {@link Account} implementation which redirects to wallet using {@link WalletConnectionRedirect} when no local key is available.
  */
-class ConnectedWalletAccountRedirect extends wallet_connection_1.ConnectedWalletAccount {
+class ConnectedWalletAccountRedirect extends wallet_connection_with_key_management_1.ConnectedWalletAccountWithKeyManagement {
     // Overriding Account methods
     /**
      * Sign a transaction by redirecting to the NEAR Wallet
