+++ src/wallet/index.ts
@@ -1,7 +1,12 @@
 
 import {
-    WalletConnectionWithKeyManagement,
+    WalletConnection,
+    RequestSignTransactionsOptions,
+    SignInOptions,
 } from './wallet-connection';
+import {
+    WalletConnectionWithKeyManagement,
+} from './wallet-connection-with-key-management';
 import {
     WalletConnectionInjected,
 } from './injected/wallet-connection-injected';
@@ -12,10 +17,20 @@ import {
     WalletConnectionRedirect,
     ConnectedWalletAccountRedirect,
 } from './redirect/wallet-connection-redirect';
-import { } from './wallet-connection-creator';
+import {
+    createWalletConnection,
+    WalletConnectionType,
+    WalletConnectionParameterOptions,
+} from './wallet-connection-creator';
 import * as wallet_injected_interface from './injected/wallet-injected-interface';
 
 export {
+    createWalletConnection,
+    WalletConnectionType,
+    WalletConnectionParameterOptions,
+    WalletConnection,
+    RequestSignTransactionsOptions,
+    SignInOptions,
     WalletConnectionWithKeyManagement,
     WalletConnectionInjected,
     WalletConnectionInjectedWithLocalFckManagement,
